# resources/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Survey API
  version: 1.0.0
  description: API Documentation for the Survey Application
components:
  schemas:
...
    SubmitResponsesRequest:
      type: object
      properties:
        responses:
          type: array
          items:
            $ref: '#/components/schemas/Response'
    Response:
      type: object
      properties:
        id:
          type: integer
        survey_id:
          type: integer
        question_id:
          type: integer
        response:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /api/surveys/{survey}/responses:
    post:
      tags:
        - Responses
      summary: Submit responses to a survey
      operationId: storeResponses
      parameters:
        - name: survey
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitResponsesRequest'
      responses:
        201:
          description: Responses submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Responses submitted successfully
...